--- rewrite-core/src/main/java/org/openrewrite/marker/OperatingSystem.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.marker;
-
-import lombok.EqualsAndHashCode;
-import lombok.Value;
-import lombok.With;
-
-import java.util.UUID;
-
-@Value
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-@With
-public class OperatingSystem implements Marker {
-    @EqualsAndHashCode.Include
-    UUID id;
-
-    Type type;
-
-    public enum Type {
-        Unix,
-        Windows
-    }
-}
