+++ rewrite-core/src/main/java/org/openrewrite/marker/OsProvenance.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.marker.ci;
+package org.openrewrite.marker;
 
 import com.sun.jna.platform.win32.Kernel32Util;
 import lombok.AllArgsConstructor;
@@ -22,7 +22,7 @@ import lombok.With;
 import org.openrewrite.Tree;
 import org.openrewrite.internal.lang.NonNull;
 import org.openrewrite.internal.lang.Nullable;
-import org.openrewrite.marker.Marker;
+import org.openrewrite.marker.ci.POSIXUtil;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -33,10 +33,10 @@ import java.util.regex.Pattern;
 import static java.util.Collections.emptyList;
 
 /**
- * Detection logic from <a href="Gradle">https://github.com/gradle/gradle/blob/master/subprojects/base-services/src/main/java/org/gradle/internal/os/OperatingSystem.java</a>
+ * Detection logic from <a href="Gradle">https://github.com/gradle/gradle/blob/master/subprojects/base-services/src/main/java/org/gradle/internal/os/OsProvenance.java</a>
  */
 @SuppressWarnings("StaticInitializerReferencesSubClass")
-public abstract class OperatingSystem implements Marker {
+public abstract class OsProvenance implements Marker {
     public static final Windows WINDOWS = new Windows();
     public static final MacOs MAC_OS = new MacOs();
     public static final Solaris SOLARIS = new Solaris();
@@ -44,19 +44,19 @@ public abstract class OperatingSystem implements Marker {
     public static final FreeBSD FREE_BSD = new FreeBSD();
     public static final Unix UNIX = new Unix();
 
-    private static OperatingSystem currentOs;
+    private static OsProvenance currentOs;
     private final String toStringValue;
     private final String osName;
     private final String osVersion;
 
-    OperatingSystem() {
+    OsProvenance() {
         osName = System.getProperty("os.name");
         osVersion = System.getProperty("os.version");
         toStringValue = getName() + " " + getVersion() + " " + System.getProperty("os.arch");
     }
 
     public static String hostname() {
-        OperatingSystem currentOs = current();
+        OsProvenance currentOs = current();
         if (currentOs.isWindows()) {
             try {
                 return Kernel32Util.getComputerName();
@@ -74,7 +74,7 @@ public abstract class OperatingSystem implements Marker {
         return "localhost";
     }
 
-    public static OperatingSystem current() {
+    public static OsProvenance current() {
         if (currentOs == null) {
             currentOs = forName(System.getProperty("os.name"));
         }
@@ -86,7 +86,7 @@ public abstract class OperatingSystem implements Marker {
         currentOs = null;
     }
 
-    public static OperatingSystem forName(String os) {
+    public static OsProvenance forName(String os) {
         String osName = os.toLowerCase();
         if (osName.contains("windows")) {
             return WINDOWS;
@@ -214,7 +214,7 @@ public abstract class OperatingSystem implements Marker {
 
     @AllArgsConstructor
     @EqualsAndHashCode(callSuper = true, onlyExplicitlyIncluded = true)
-    static class Windows extends OperatingSystem {
+    static class Windows extends OsProvenance {
         String nativePrefix;
 
         @With
@@ -313,7 +313,7 @@ public abstract class OperatingSystem implements Marker {
 
     @AllArgsConstructor
     @EqualsAndHashCode(callSuper = true, onlyExplicitlyIncluded = true)
-    static class Unix extends OperatingSystem {
+    static class Unix extends OsProvenance {
         String nativePrefix;
 
         @With
